requirements.txt
click==8.1.7
markdown-it-py==3.0.0
mdurl==0.1.2
pillow==10.4.0
Pygments==2.18.0
rich==13.9.1
